+++ physics/CHANGELOG.md
@@ -12,6 +12,7 @@ changes will be documented in this file.
 
 - Fixed panic in wire baseline calibration for new runs by removing "safeguards".
 - Fixed panic in wire gain calibration for new runs by removing "safeguards".
+- Fixed panic in wire delay calibration for new runs by removing "safeguards".
 
 <!-- next-url -->
 [Unreleased]: https://github.com/ALPHA-g-Experiment/alpha-g/compare/alpha_g_physics-v0.1.2...HEAD